Refactored traits into separate modules.
